MET-4285 Truncate long decimals geo, re-organize solr properties
